ChangeMethodTargetToStatic from Is.is to Matchers.is first
