+++ src/main/resources/META-INF/rewrite/hamcrest.yml
@@ -39,7 +39,12 @@ tags:
   - hamcrest
   - assertj
 recipeList:
-  # First remove wrapping `is(Matcher)` calls such that further recipes will match
+  # First change `is(..)` to `Matchers.is(..)` for consist
+  - org.openrewrite.java.ChangeMethodTargetToStatic:
+      methodPattern: org.hamcrest.core.Is is(..)
+      fullyQualifiedTargetTypeName: org.hamcrest.Matchers
+
+  # Then remove wrapping `is(Matcher)` calls such that further recipes will match
   - org.openrewrite.java.testing.hamcrest.RemoveIsMatcher
 
   # Then remove calls to `MatcherAssert.assertThat(String, is(Matcher))`
